Move features: changeTheme to entity/theme, product/sortBy to category page
