+++ src/entities/theme/index.ts
@@ -1,2 +1,3 @@
 export { themeSlice } from './model/slice'
 export { ThemeProvider } from './lib/ThemeProvider'
+export { ChangeTheme } from './ui/ChangeTheme/ChangeTheme'
