--- src/features/product/sortBy/index.ts
@@ -1,2 +1,0 @@
-export { SortByDropdown } from './ui/SortByDropdown/SortByDropdown'
-export type { ProductSortBy } from './model/types'
