--- src/features/theme/changeTheme/index.ts
@@ -1,1 +1,0 @@
-export { ChangeTheme } from './ui/ChangeTheme/ChangeTheme'
