+++ src/pages/category/ui/Page/Page.tsx
@@ -4,12 +4,13 @@ import { z } from 'zod'
 import { useCategoryDetailsQuery } from '@/entities/category'
 import type { CategoryId } from '@/entities/category/model/types'
 import { useFeatureToggle } from '@/entities/featureToggle'
-import { type ProductSortBy, SortByDropdown } from '@/features/product/sortBy'
 import { useTypedParams, useTypedQueryParams } from '@/shared/lib/router'
 import { useAppDispatch, useAppSelector } from '@/shared/redux'
 import { PageHeader } from '@/shared/ui'
 import { BaseProductList } from '@/widgets/BaseProductList'
 import { categoryPageSlice } from '../../model/slice'
+import type { ProductSortBy } from '../../model/types'
+import { SortByDropdown } from '../SortByDropdown/SortByDropdown'
 
 const pageParamsSchema = z.object({
   categoryId: z.coerce
