+++ src/widgets/LayoutHeader/ui/LayoutHeader/LayoutHeader.tsx
@@ -3,8 +3,8 @@ import type { ReactNode } from 'react'
 import { useCartQuery } from '@/entities/cart/api/cartApi'
 import { useFeatureToggle } from '@/entities/featureToggle'
 import { selectIsAuthorized } from '@/entities/session'
+import { ChangeTheme } from '@/entities/theme'
 import { useWishlistProductsQuery } from '@/entities/wishlist'
-import { ChangeTheme } from '@/features/theme/changeTheme'
 import { useAppSelector } from '@/shared/redux'
 import { Logo } from '../Logo/Logo'
 import css from './LayoutHeader.module.css'
